Image is now pulled from config map data
