+++ frontend/packages/kubevirt-plugin/src/components/create-vm-wizard/redux/initial-state/storage-tab-initial-state.ts
@@ -1,6 +1,6 @@
 import { ConfigMapKind } from '@console/internal/module/k8s';
 
-import { VM_TEMPLATE_NAME_PARAMETER, WINTOOLS_CONTAINER_NAMES } from '../../../../constants';
+import { VM_TEMPLATE_NAME_PARAMETER } from '../../../../constants';
 import {
   DUMMY_VM_NAME,
   ROOT_DISK_NAME,
@@ -14,6 +14,7 @@ import {
   DiskType,
   VolumeType,
 } from '../../../../constants/vm/storage';
+import { winToolsContainerNames } from '../../../../constants/vm/wintools';
 import { DataVolumeWrapper } from '../../../../k8s/wrapper/vm/data-volume-wrapper';
 import { DiskWrapper } from '../../../../k8s/wrapper/vm/disk-wrapper';
 import { VolumeWrapper } from '../../../../k8s/wrapper/vm/volume-wrapper';
@@ -91,6 +92,8 @@ const getContainerStorage = (
   };
 };
 
+const containerNames = winToolsContainerNames();
+
 export const windowsToolsStorage: VMWizardStorage = {
   type: VMWizardStorageType.WINDOWS_GUEST_TOOLS,
   disk: DiskWrapper.initializeFromSimpleData({
@@ -102,7 +105,7 @@ export const windowsToolsStorage: VMWizardStorage = {
     name: WINTOOLS_DISK_NAME,
     type: VolumeType.CONTAINER_DISK,
     typeData: {
-      image: WINTOOLS_CONTAINER_NAMES[window.SERVER_FLAGS.branding] || WINTOOLS_CONTAINER_NAMES.okd,
+      image: containerNames[window.SERVER_FLAGS.branding] || containerNames?.okd,
     },
   }).asResource(),
 };
