+++ frontend/packages/kubevirt-plugin/src/constants/vm/constants.ts
@@ -66,33 +66,7 @@ export const VM_DETAIL_EVENTS_HREF = 'events';
 
 export const DUMMY_VM_NAME = 'vm';
 
-// NOTE(yaacov):
-// BZ 1944273, wintool container version should matches the "latest" image at the time of the release
-// Bug 1942839, container images should use digest for competability with disconnected enviorments
-//
-// To get the image's digest you could the following steps:
-//
-// * Step 1:
-// skopeo login registry.redhat.io
-// Login Succeeded!
-//
-// * Step 2: skopeo inspect docker://registry.redhat.io/container-native-virtualization/virtio-win:v2.6.0 | jq '.Digest'
-// "sha256:011060472f068e42e2c0c0b3451a99b5607dd037ba70945004f98b2de74b89a2"
-//
-// * Step 3:
-// podman pull --authfile /root/ocp4-disconnected/pull-secret.json registry.redhat.io/container-native-virtualization/virtio-win@sha256:<digest>
-export const WINTOOLS_CONTAINER_DIGEST =
-  '011060472f068e42e2c0c0b3451a99b5607dd037ba70945004f98b2de74b89a2';
-export const WINTOOLS_CONTAINER_NAME_DOWNSTREAM = `registry.redhat.io/container-native-virtualization/virtio-win@sha256:${WINTOOLS_CONTAINER_DIGEST}`;
-
-export const WINTOOLS_CONTAINER_NAMES = {
-  openshift: WINTOOLS_CONTAINER_NAME_DOWNSTREAM,
-  ocp: WINTOOLS_CONTAINER_NAME_DOWNSTREAM,
-  online: WINTOOLS_CONTAINER_NAME_DOWNSTREAM,
-  dedicated: WINTOOLS_CONTAINER_NAME_DOWNSTREAM,
-  azure: WINTOOLS_CONTAINER_NAME_DOWNSTREAM,
-  okd: 'kubevirt/virtio-container-disk', // comunity version is always "latest"
-};
+export const VIRTIO_WIN_IMAGE = 'virtio-win-image';
 
 export const PENDING_RESTART_LABEL = '(pending restart)';
 export const getPVCUploadURL = (pvcNamespace: string): string =>
