+++ frontend/packages/kubevirt-plugin/src/selectors/vm/volume.ts
@@ -1,6 +1,6 @@
 import * as _ from 'lodash';
 
-import { WINTOOLS_CONTAINER_NAMES } from '../../constants';
+import { winToolsContainerNames } from '../../constants/vm/wintools';
 
 export const getVolumePersistentVolumeClaimName = (volume) =>
   _.get(volume, 'persistentVolumeClaim.claimName');
@@ -11,5 +11,7 @@ export const getVolumeCloudInitNoCloud = (volume) => volume && volume.cloudInitN
 export const getVolumeContainerImage = (volume) =>
   volume && volume.containerDisk && volume.containerDisk.image;
 
-export const isWinToolsImage = (image) =>
-  Object.values(WINTOOLS_CONTAINER_NAMES).find((winTool) => image && image.startsWith(winTool));
+export const isWinToolsImage = (image) => {
+  const containerNames = winToolsContainerNames();
+  return Object.values(containerNames).find((winTool) => image && image.startsWith(winTool));
+};
